requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.1
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.76.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.6.8
passlib==1.7.4
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pypika-tortoise==0.1.4
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.18.0
tortoise-orm==0.19.0
typing-extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
